consumer/requirements.txt
asyncio==3.4.3
pika==1.3.2
PyYAML==6.0.2
websockets==14.1
